app/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
google-cloud-firestore==2.15.0
openai==1.30.5
python-dotenv==1.0.1    